refactor(core): Move callModuleInitHook into dedicated file
